Rename parseReferences.js to findReferences.js
This is consistent with the single export (findReferences.js now exports
`findReferences`)
