+++ src/plugins/github/findReferences.js
